Enhance UptimeKuma with Socket-Proxy option
UptimeKuma settings has a Docker Hosts section. This enhancement lets the user connect to tcp://socket-proxy:2375 thus allowing the UptimeKuma monitor to watch the container directly in docker.

This commit/enhancement gives user an option to link to socket-proxy container.  The compose will ask if a socket-proxy should be used.  
If yes, will then ask if one already exists.  
- If yes, link to the existing socket-proxy.  
- If no, create and link to it.