Update Mealie cosmos-compose.json (#188)
Current Servapp is pulling v1.0.0-RC2 which is over a year old and missing out on updates due to the RC2 argument. I can't find any documentation what RC2 stands for, or if it's necessary for the Cosmos configuration, but the latest version v2.8.0 seems to be working properly on my machine.